+++ src/basilisp/cli.py
@@ -79,17 +79,22 @@ def bootstrap_repl(ctx: compiler.CompilerContext, which_ns: str) -> types.Module
     return importlib.import_module(REPL_NS)
 
 
-def init_path(args: argparse.Namespace) -> None:
-    def append_once(path: str) -> None:
-        if path in sys.path:
-            return
-        sys.path.insert(0, path)
+def _sys_path_prepend_once(path: str) -> None:
+    """Prepend an entry to `sys.path` if it is not already in `sys.path`."""
+    if path in sys.path:
+        return
+    sys.path.insert(0, path)
 
+
+def init_path(args: argparse.Namespace, unsafe_path: str = "") -> None:
+    """Prepend any import group arguments to `sys.path`, including `unsafe_path` (which
+    defaults to the empty string) if --include-unsafe-path is specified."""
     for path in args.include_path or []:
         p = pathlib.Path(path).resolve()
-        append_once(str(p))
+        _sys_path_prepend_once(str(p))
 
-    append_once("")
+    if args.include_unsafe_path:
+        _sys_path_prepend_once(unsafe_path)
 
 
 def _to_bool(v: Optional[str]) -> Optional[bool]:
@@ -280,12 +285,35 @@ def _add_debug_arg_group(parser: argparse.ArgumentParser) -> None:
 
 
 def _add_import_arg_group(parser: argparse.ArgumentParser) -> None:
-    group = parser.add_argument_group("path options")
+    group = parser.add_argument_group(
+        "path options",
+        description=(
+            "The path options below can be used to control how Basilisp (and Python) "
+            "find your code."
+        ),
+    )
+    group.add_argument(
+        "--include-unsafe-path",
+        action=_set_envvar_action(
+            "BASILISP_INCLUDE_UNSAFE_PATH", parent=argparse._StoreAction
+        ),
+        nargs="?",
+        const=True,
+        type=_to_bool,
+        help=(
+            "if true, automatically prepend a potentially unsafe path to `sys.path`; "
+            "this is the Basilisp equivalent to the PYTHONSAFEPATH environment variable "
+            "(env: BASILISP_INCLUDE_UNSAFE_PATH; default: true)"
+        ),
+    )
     group.add_argument(
         "-p",
         "--include-path",
         action="append",
-        help="path to add to `sys.path`; maybe specified more than once",
+        help=(
+            "path to prepend to `sys.path`; may be specified more than once to "
+            "include multiple paths (env: PYTHONPATH)"
+        ),
     )
 
 
@@ -305,8 +333,8 @@ def _add_runtime_arg_group(parser: argparse.ArgumentParser) -> None:
         const=_to_bool(os.getenv("BASILISP_USE_DATA_READERS_ENTRY_POINT", "true")),
         type=_to_bool,
         help=(
-            "If true, Load data readers from importlib entry points in the "
-            '"basilisp_data_readers" group. (env: '
+            "if true, Load data readers from importlib entry points in the "
+            '"basilisp_data_readers" group (env: '
             "BASILISP_USE_DATA_READERS_ENTRY_POINT; default: true)"
         ),
     )
@@ -560,7 +588,6 @@ def run(
 
     opts = _compiler_opts(args)
     basilisp.init(opts)
-    init_path(args)
     ctx = compiler.CompilerContext(
         filename=(
             CLI_INPUT_FILE_PATH
@@ -583,6 +610,7 @@ def run(
             cli_args_var.bind_root(vec.vector(args.args))
 
         if args.code:
+            init_path(args)
             eval_str(target, ctx, ns, eof)
         elif args.load_namespace:
             # Set the requested namespace as the *main-ns*
@@ -590,10 +618,13 @@ def run(
             assert main_ns_var is not None
             main_ns_var.bind_root(sym.symbol(target))
 
+            init_path(args)
             importlib.import_module(munge(target))
         elif target == STDIN_FILE_NAME:
+            init_path(args)
             eval_stream(io.TextIOWrapper(sys.stdin.buffer, encoding="utf-8"), ctx, ns)
         else:
+            init_path(args, unsafe_path=str(pathlib.Path(target).resolve()))
             eval_file(target, ctx, ns)